Rebuild ActiveSupport automatic loading of files
Define some core extensions needed for the automatic loading of Rails
files. In Rails everything under the `/app` folder is autoloaded when
needed so that we never have to call `require`. For example, in our
test app the first time we call for the `Post` model, the proper file
will be required for us.